+++ pkgs/test_core/CHANGELOG.md
@@ -5,8 +5,10 @@
 * Respect `NO_COLOR`, `CLICOLOR`, `FORCE_COLOR`, `CLICOLOR_FORCE`, and `TERM=dumb`
   environment variables for color output detection.
 * Fix a bug where `-c exe` tests would hang on exit on windows.
-* In `GithubReporter`, group contiguous passing tests into a collapsible group 
-  to reduce log noise in GitHub Actions.
+* `GithubReporter`:
+    * Group contiguous passing and skipped tests into collapsible groups to
+      reduce log noise in GitHub Actions.
+    * Updated skipped icon to ⚠️.
 
 ## 0.6.17